Fix out of range because of float cast (#1753)

+++ sensing/pointcloud_preprocessor/src/ground_filter/scan_ground_filter_nodelet.cpp
@@ -85,8 +85,7 @@ void ScanGroundFilterComponent::convertPointcloud(
 
   for (size_t i = 0; i < in_cloud->points.size(); ++i) {
     auto radius{static_cast<float>(std::hypot(in_cloud->points[i].x, in_cloud->points[i].y))};
-    auto theta{static_cast<float>(
-      normalizeRadian(std::atan2(in_cloud->points[i].x, in_cloud->points[i].y), 0.0))};
+    auto theta{normalizeRadian(std::atan2(in_cloud->points[i].x, in_cloud->points[i].y), 0.0)};
     auto radial_div{static_cast<size_t>(std::floor(theta / radial_divider_angle_rad_))};
 
     current_point.radius = radius;
